+++ tw-tasks-core/src/main/java/com/transferwise/tasks/health/ITasksStateMonitor.java
@@ -1,10 +1,12 @@
 package com.transferwise.tasks.health;
 
+import com.transferwise.tasks.domain.TaskStatus;
 import java.util.Map;
+import org.apache.commons.lang3.tuple.Pair;
 
 public interface ITasksStateMonitor {
 
   Map<String, Integer> getErroneousTasksCountByType();
 
-  Map<String, Integer> getStuckTasksCountByType();
+  Map<Pair<TaskStatus, String>, Integer> getStuckTasksCountByType();
 }